ref: rename elements subfolder to components
idk why i did that, must have wanted to be edgy
