+++ frontend/src/components/Dropdown/components/SelectCombobox/ComboboxClearButton.tsx
