+++ frontend/src/components/Dropdown/components/SelectCombobox/LabelIcon.tsx
